Refactor parse_bicep to use it in both aci policy gen and VM script generation
In preparation for vn2 yaml generation
